pyproject.toml
[project]
name = "pytorch-lightning-hydra-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=24.8.0",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "lightning>=2.4.0",
    "loguru>=0.7.2",
    "matplotlib>=3.9.2",
    "numpy==1.26.0",
    "pytest>=8.3.3",
    "rich>=13.9.2",
    "rootutils>=1.0.7",
    "seaborn>=0.13.2",
    "tensorboard>=2.18.0",
    "timm>=1.0.9",
    "torch==2.2.2",
    "torchmetrics>=1.4.2",
    "torchvision>=0.17.2",
    "typing-extensions>=4.12.2",
]
